summary

+++ werr/werr/UI/UmsjonRepo.cpp
@@ -21,7 +21,7 @@ void UmsjonRepo::storeAllToppings(const vector<Topping> &toppings) {
 
     int toppingcount = toppings.size();
 
-    fout.write((char*)(&toppingcount), sizeof(double));
+    fout.write((char*)(&toppingcount), sizeof(int));
 
     for (int i=0; i<toppingcount; i++) {
         toppings[i].write(fout);
@@ -38,7 +38,7 @@ vector<Topping> UmsjonRepo::retriveAllToppings() {
     if(fin.is_open()) {
         int toppingcount;
 
-        fin.read((char*)(&toppingcount), sizeof(double));
+        fin.read((char*)(&toppingcount), sizeof(int));
 
         for (int i=0; i<toppingcount; i++) {
             topping.read(fin);
